requirements.txt
fastapi==0.135.2
python-dotenv
uvicorn==0.42.0
pydantic==2.12.5
anyio==4.13.0
starlette==1.0.0
python-jose[cryptography]==3.3.0
supabase

# Test dependencies
pytest
pytest-asyncio
httpx
hypothesis
